Update to 1.10, add JEI support

+++ src/main/java/zmaster587/libVulpes/util/CreativeBattery.java
@@ -13,6 +13,6 @@ public class CreativeBattery extends UniversalBattery {
 	}
 	@Override
 	public int getMaxEnergyStored() {
-		return Integer.MAX_VALUE;
+		return Short.MAX_VALUE;
 	}
 }
